requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
pandas==2.0.3
pyyaml==6.0.1
tabulate==0.9.0
colorama==0.4.6
python-dotenv==1.0.0